+++ activerecord/lib/active_record/associations/association_scope.rb
@@ -5,7 +5,7 @@ module ActiveRecord
 
       attr_reader :association, :alias_tracker
 
-      delegate :klass, :owner, :reflection, :interpolate, :to => :association
+      delegate :klass, :owner, :reflection, :to => :association
       delegate :chain, :scope_chain, :options, :source_options, :active_record, :to => :reflection
 
       def initialize(association)