+++ activerecord/lib/active_record/associations/has_and_belongs_to_many_association.rb
@@ -18,16 +18,12 @@ module ActiveRecord
           end
         end
 
-        if options[:insert_sql]
-          owner.connection.insert(interpolate(options[:insert_sql], record))
-        else
-          stmt = join_table.compile_insert(
-            join_table[reflection.foreign_key]             => owner.id,
-            join_table[reflection.association_foreign_key] => record.id
-          )
+        stmt = join_table.compile_insert(
+          join_table[reflection.foreign_key]             => owner.id,
+          join_table[reflection.association_foreign_key] => record.id
+        )
 
-          owner.connection.insert stmt
-        end
+        owner.connection.insert stmt
 
         record
       end
@@ -39,22 +35,17 @@ module ActiveRecord
         end
 
         def delete_records(records, method)
-          if sql = options[:delete_sql]
-            records = load_target if records == :all
-            records.each { |record| owner.connection.delete(interpolate(sql, record)) }
-          else
-            relation  = join_table
-            condition = relation[reflection.foreign_key].eq(owner.id)
+          relation  = join_table
+          condition = relation[reflection.foreign_key].eq(owner.id)
 
-            unless records == :all
-              condition = condition.and(
-                relation[reflection.association_foreign_key]
-                  .in(records.map { |x| x.id }.compact)
-              )
-            end
-
-            owner.connection.delete(relation.where(condition).compile_delete)
+          unless records == :all
+            condition = condition.and(
+              relation[reflection.association_foreign_key]
+                .in(records.map { |x| x.id }.compact)
+            )
           end
+
+          owner.connection.delete(relation.where(condition).compile_delete)
         end
 
         def invertible_for?(record)