+++ activerecord/test/models/company.rb
@@ -36,9 +36,7 @@ module Namespaced
 end
 
 class Firm < Company
-  has_many :clients, -> { order "id" }, :dependent => :destroy, :counter_sql =>
-      "SELECT COUNT(*) FROM companies WHERE firm_id = 1 " +
-      "AND (#{QUOTED_TYPE} = 'Client' OR #{QUOTED_TYPE} = 'SpecialClient' OR #{QUOTED_TYPE} = 'VerySpecialClient' )",
+  has_many :clients, -> { order "id" }, :dependent => :destroy,
       :before_remove => :log_before_remove,
       :after_remove  => :log_after_remove
   has_many :unsorted_clients, :class_name => "Client"
@@ -53,17 +51,6 @@ class Firm < Company
   has_many :clients_with_interpolated_conditions, ->(firm) { where "rating > #{firm.rating}" }, :class_name => "Client"
   has_many :clients_like_ms, -> { where("name = 'Microsoft'").order("id") }, :class_name => "Client"
   has_many :clients_like_ms_with_hash_conditions, -> { where(:name => 'Microsoft').order("id") }, :class_name => "Client"
-  has_many :clients_using_sql, :class_name => "Client", :finder_sql => proc { "SELECT * FROM companies WHERE client_of = #{id}" }
-  has_many :clients_using_counter_sql, :class_name => "Client",
-           :finder_sql  => proc { "SELECT * FROM companies WHERE client_of = #{id} " },
-           :counter_sql => proc { "SELECT COUNT(*) FROM companies WHERE client_of = #{id}" }
-  has_many :clients_using_zero_counter_sql, :class_name => "Client",
-           :finder_sql  => proc { "SELECT * FROM companies WHERE client_of = #{id}" },
-           :counter_sql => proc { "SELECT 0 FROM companies WHERE client_of = #{id}" }
-  has_many :no_clients_using_counter_sql, :class_name => "Client",
-           :finder_sql  => 'SELECT * FROM companies WHERE client_of = 1000',
-           :counter_sql => 'SELECT COUNT(*) FROM companies WHERE client_of = 1000'
-  has_many :clients_using_finder_sql, :class_name => "Client", :finder_sql => 'SELECT * FROM companies WHERE 1=1'
   has_many :plain_clients, :class_name => 'Client'
   has_many :readonly_clients, -> { readonly }, :class_name => 'Client'
   has_many :clients_using_primary_key, :class_name => 'Client',