+++ activerecord/CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Rails 4.0.0 (unreleased) ##
 
+*   Removed `:finder_sql` and `:counter_sql` collection association options. Please
+    use scopes instead.
+
+    *Jon Leighton*
+
+*   Removed `:insert_sql` and `:delete_sql` `has_and_belongs_to_many`
+    association options. Please use `has_many :through` instead.
+
+    *Jon Leighton*
+
 *   The migration generator now creates a join table with (commented) indexes every time
     the migration name contains the word `join_table`: